refactor: Namespace as `my/default_reference_controller`
