+++ spec/controllers/my/default_reference_controller_spec.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
 
-describe DefaultReferencesController do
+describe My::DefaultReferenceController do
   describe "forbidden actions" do
     context "when not signed in" do
-      specify { expect(post(:update, params: { id: 1 })).to redirect_to_signin_form }
+      specify { expect(put(:update, params: { id: 1 })).to redirect_to_signin_form }
     end
   end
 
@@ -16,7 +16,7 @@ describe DefaultReferencesController do
 
     it "calls `DefaultReference.set`" do
       expect(DefaultReference).to receive(:set).with(session, reference).and_call_original
-      post :update, params: { id: reference.id }
+      put :update, params: { id: reference.id }
     end
   end
 end
